Cleanup current fhir return handler
